Feat(Auth):Implement authentication and user management features with registration, login, and profile functionalities

+++ initial/app/Http/Controllers/Studentcontroller.php
@@ -5,29 +5,42 @@ namespace App\Http\Controllers;
 use App\Models\Majors;
 use App\Models\Student;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class StudentController extends Controller
 {
     public function index()
     {
+        if(! Gate::allows('view-student') ){
+            abort(403, 'Unauthorized action.');
+        }
         $students = Student::with('majors')->get();
         return view('students.index', compact('students'));
     }
 
     public function show(string $id)
     {
+        if (!Gate::allows('view-student')) {
+            abort(403, 'Unauthorized action.');
+        }
         $student = Student::with('majors')->find($id);
         return view('students.view', compact('student'));
     }
 
     public function create()
     {
+        if (!Gate::allows('store-student')) {
+            abort(403, 'Unauthorized action.');
+        }
         $majors = Majors::all();
         return view('students.create', compact('majors'));
     }
 
     public function store(Request $request)
     {
+        if (!Gate::allows('store-student')) {
+            abort(403, 'Unauthorized action.');
+        }
         $validated = $request->validate([
             'name' => 'required',
             'student_id_number' => 'required|unique:students|max:9',
@@ -55,6 +68,9 @@ class StudentController extends Controller
 
     public function edit(string $id)
     {
+        if (!Gate::allows('edit-student')) {
+            abort(403, 'Unauthorized action.');
+        }
         $student = Student::with('majors')->find($id);
         $majors = Majors::all();
         return view('students.edit', compact('student', 'majors'));
@@ -62,6 +78,9 @@ class StudentController extends Controller
 
     public function update(Request $request, string $id)
     {
+        if (!Gate::allows('edit-student')) {
+            abort(403, 'Unauthorized action.');
+        }
         $validated = $request->validate([
             'name' => 'required',
             'student_id_number' => "required|unique:students,student_id_number,$id|max:9",
@@ -91,6 +110,9 @@ class StudentController extends Controller
 
     public function destroy(string $id)
     {
+        if (!Gate::allows('destroy-student')) {
+            abort(403, 'Unauthorized action.');
+        }
         $student = Student::findOrFail($id);
         $student->delete();
